Move class AddCoverArtToAlbumsRunnable from com.restmash.app to com.restmash.musicbrainz
